requirements.txt
alembic==1.15.1
bcrypt==4.3.0
fastapi==0.115.11
python-jose==3.4.0
sqlalchemy==2.0.39
asyncpg==0.30.0
uvicorn==0.34.0

sqlmodel==0.0.24
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic_settings==2.8.1
python-multipart==0.0.20
